+++ glpi_data/plugins/ativawallpaper/front/history.php
@@ -42,10 +42,8 @@ TemplateRenderer::getInstance()->display('@ativawallpaper/history.html.twig', [
     'urls'       => [
         'action'    => $base . '/front/action.php',
         'dashboard' => $base . '/front/dashboard.php',
-        'glpi_agent' => $base . '/front/glpi_agent.php',
         'history'   => $base . '/front/history.php',
         'events'    => $base . '/front/events.php',
-        'updates'   => $base . '/front/updates.php',
         'settings'  => $base . '/front/settings.php',
         'image'     => $base . '/front/image.php',
     ],