+++ glpi_data/plugins/ativawallpaper/front/settings.php
@@ -48,10 +48,8 @@ TemplateRenderer::getInstance()->display('@ativawallpaper/settings.html.twig', [
     'urls'              => [
         'action'    => $base . '/front/action.php',
         'dashboard' => $base . '/front/dashboard.php',
-        'glpi_agent' => $base . '/front/glpi_agent.php',
         'history'   => $base . '/front/history.php',
         'events'    => $base . '/front/events.php',
-        'updates'   => $base . '/front/updates.php',
         'settings'  => $base . '/front/settings.php',
     ],
 ]);